chore(deps): update outsystems design tokens package (#30586)
Updating outsystems-design-tokens reference. This package removes the
Inter font from being the default font, and replaced it with the
respective system fonts.

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying. -->
The default font was Inter, that had to be loaded from Google CDN,
causing CSP problems with default rules.

## What is the new behavior?
<!-- Please describe the behavior or changes that are being added by
this PR. -->
- Update to latest version of `outsystems-design-tokens` that no longer
uses the Inter font.

## Does this introduce a breaking change?
- [ ] Yes
- [x] No

<!--
  If this introduces a breaking change:
1. Describe the impact and migration path for existing applications
below.
  2. Update the BREAKING.md file with the breaking change.
3. Add "BREAKING CHANGE: [...]" to the commit description when merging.
See
https://github.com/ionic-team/ionic-framework/blob/main/docs/CONTRIBUTING.md#footer
for more information.
-->


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->